# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Peexoo's API swagger documentation",
        "description": "This is a sample Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Logs a user in",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "User Oject",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/auth/create": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Creates a User",
                "operationId": "createAccount",
                "requestBody": {
                    "description": "User Oject",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/auth/verify/{email}/{pin}": {
            "patch": {
                "tags": [
                    "auth"
                ],
                "summary": "Verifies a User's account",
                "operationId": "verifyAccount",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "The email for the verification sent to the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pin",
                        "in": "path",
                        "description": "The verification code sent to the user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/auth/forgotPassword/{url_path}/{email}": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Send a password reset Link",
                "operationId": "forgotPassword",
                "parameters": [
                    {
                        "name": "url_path",
                        "in": "path",
                        "description": "The Url to the page that would conclude the reset. The final url would be {url_path}/token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email of the user requesting for email reset",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/auth/resetpassword": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "",
                "operationId": "resetPassword",
                "requestBody": {
                    "description": "User Oject",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetForgotPassword"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/auth/photographer/create": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Creates a Photographer User account",
                "operationId": "createQuickAccountForPhotographer",
                "requestBody": {
                    "description": "User Oject",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ResetForgotPassword": {
                "title": "ResetForgotPassword Object Representation",
                "description": "ResetForgotPassword model",
                "required": [
                    "user_id",
                    "password"
                ],
                "properties": {
                    "email": {
                        "title": "Email",
                        "description": "",
                        "type": "string"
                    },
                    "resetToken": {
                        "title": "Reset Token",
                        "description": "",
                        "type": "string"
                    },
                    "new_password": {
                        "title": "New Password",
                        "description": "",
                        "type": "string"
                    },
                    "repeat_new_password": {
                        "title": "Repeat New Password",
                        "description": "",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ResetForgotPassword"
                }
            },
            "User": {
                "title": "User Object Representation",
                "description": "User model",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "title": "Id",
                        "description": "User Id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "last_name": {
                        "title": "Name",
                        "description": "User Last Name",
                        "type": "string"
                    },
                    "first_name": {
                        "title": "Name",
                        "description": "User First Name",
                        "type": "string"
                    },
                    "email": {
                        "title": "Email",
                        "description": "User Email",
                        "type": "string"
                    },
                    "tel_number": {
                        "title": "Tel Number",
                        "description": "User Tel number",
                        "type": "string"
                    },
                    "password": {
                        "title": "password",
                        "description": "User password",
                        "type": "string",
                        "format": "password"
                    },
                    "profile_image": {
                        "title": "Profile Picture",
                        "description": "User Profile Picture. This is a url to the image. ",
                        "type": "string"
                    },
                    "job_description": {
                        "title": "Job Description",
                        "description": "The person's Job Description in a word!!",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
...
        },
        "securitySchemes": {
            "api_key_scheme": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}